fix(dialog): reuse getInstance methods
